[Red] Add App test useSelector / useDispatch
